Upgrade PlantUML to 1.2020.12 (#273)
* Upgrade PlantUML to 1.2020.12
* Remove a dependency on PlantUML FileUtils class for a simple file copy

+++ src/plantuml-asl/src/net/sourceforge/plantuml/nwdiag/NwDiagram.java
@@ -44,7 +44,6 @@ import java.util.regex.Pattern;
 
 import net.sourceforge.plantuml.AnnotatedWorker;
 import net.sourceforge.plantuml.ColorParam;
-import net.sourceforge.plantuml.Dimension2DDouble;
 import net.sourceforge.plantuml.FileFormatOption;
 import net.sourceforge.plantuml.ISkinParam;
 import net.sourceforge.plantuml.Scale;
@@ -192,8 +191,8 @@ public class NwDiagram extends UmlDiagram {
 			margin1 = 0;
 			margin2 = 0;
 		}
-		final ImageBuilder imageBuilder = ImageBuilder.buildB(new ColorMapperIdentity(), false, ClockwiseTopRightBottomLeft.margin1margin2((double) margin1, (double) margin2),
-		null, "", "", dpiFactor, null);
+		final ImageBuilder imageBuilder = ImageBuilder.buildB(new ColorMapperIdentity(), false,
+				ClockwiseTopRightBottomLeft.margin1margin2(margin1, margin2), null, "", "", dpiFactor, null);
 		TextBlock result = getTextBlock();
 		result = new AnnotatedWorker(this, skinParam, fileFormatOption.getDefaultStringBounder()).addAdd(result);
 		imageBuilder.setUDrawable(result);
